// functions/src/models/oldSQLDB.model.ts
export interface sqlUser {
    "usrUsername": string;
    "usrEmailAddress": string;
    "usrPassword": string;
    "usrName": string;
    "usrSurname": string;
    "usrStudentNumber": string;
    "usrFirstYearYear": string;
    "usrIsSemi": string;
    "usrVerified": string;
    "usrIsHK": string;
    "tblBedieningTable_talID": string;
}